Change weight into a float so that we can take into account the nesting level when computing it. Right now the allocator uses:
    w = sum_over_defs_uses( 10 ^ nesting level );

llvm-svn: 10569

+++ llvm/include/llvm/CodeGen/LiveIntervals.h
@@ -38,8 +38,9 @@ namespace llvm {
             typedef std::pair<unsigned, unsigned> Range;
             typedef std::vector<Range> Ranges;
             unsigned reg;   // the register of this interval
-            unsigned weight; // weight of this interval (number of uses)
-            Ranges ranges; // the ranges this register is valid
+            float weight;   // weight of this interval (number of uses
+                            // * 10^loopDepth)
+            Ranges ranges;  // the ranges this register is valid
 
             Interval(unsigned r);
